move explore tooltip for single location
